requirements.txt
streamlit
requests
PyYAML
julep
pandas
geopy
folium
reportlab
